Check in Roger's changes
I found some ways of correcting the wrong temperature increase. I
re-used the NASAbasedSurfaceTemperature - using the temperature
distribution of the last timeslice based on NASA. This functions now
such that the temperature on land is just a little higher than on
oceans. But it happens that the temperature close to continents still
increases in hot spots because I have not used an interpolation method
to close the gaps. Simon's method will do that. So, I am optimistic to
see better temperature results.